cli/requirements.txt
PyYaml >= 6.0, < 7.0
pyjwt >= 2.6.0, < 2.7
